1、crowdBatchTaskPending 线程池 设置核心线程空闲 回收策略以及时间 2、线程池配置优化 3、优雅关闭非Spring管理的线程池（停机减少数据丢失） 4、优化read.me

+++ austin-web/src/main/java/com/java3y/austin/web/service/impl/MessageTemplateServiceImpl.java
@@ -70,7 +70,7 @@ public class MessageTemplateServiceImpl implements MessageTemplateService {
         Iterable<MessageTemplate> messageTemplates = messageTemplateDao.findAllById(ids);
         messageTemplates.forEach(messageTemplate -> messageTemplate.setIsDeleted(AustinConstant.TRUE));
         for (MessageTemplate messageTemplate : messageTemplates) {
-            if (messageTemplate.getCronTaskId() > 0) {
+            if (messageTemplate.getCronTaskId()!=null && messageTemplate.getCronTaskId() > 0) {
                 cronTaskService.deleteCronTask(messageTemplate.getCronTaskId());
             }
         }
